Dodana ogranicenja

+++ app/src/main/java/com/example/barrycards/drugaGodina/ActivityDrugaGodina.kt
@@ -24,17 +24,17 @@ class ActivityDrugaGodina : AppCompatActivity() {
         setContentView(R.layout.activity_druga_godina)
         val buttonDodajKolegiji = findViewById<FloatingActionButton>(R.id.button_add_kolegiji)
 
-        //val uloga = intent.getStringExtra("Uloga")
+        val uloga = intent.getStringExtra("Uloga")
 
-        //if ("Admin" == uloga || "Superuser" == uloga) {
+        if ("Admin" == uloga) {
         buttonDodajKolegiji.visibility = View.VISIBLE
         buttonDodajKolegiji.setOnClickListener {
             val intent = Intent(this@ActivityDrugaGodina, AddEditKolegiji::class.java)
             startActivityForResult(intent, ADD_DRUGI_KOLEGIJI_REQUEST)
         }
-        /*} else {
-            buttonDodajZadatak.visibility = View.GONE
-        }*/
+        } else {
+            buttonDodajKolegiji.visibility = View.GONE
+        }
 
 
 
@@ -50,7 +50,7 @@ class ActivityDrugaGodina : AppCompatActivity() {
         ) { drugiKolegiji -> adapterDrugaGodina.setDrugiKolegijiList(drugiKolegiji as List<DrugaGodinaKolegiji>) }
 
 
-        // if ("Admin" == uloga) {
+        if ("Admin" == uloga) {
         ItemTouchHelper(object : ItemTouchHelper.SimpleCallback(
             0,
             ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
@@ -83,7 +83,7 @@ class ActivityDrugaGodina : AppCompatActivity() {
 
             }
         })
-        //}
+        }
         supportActionBar!!.setDisplayHomeAsUpEnabled(true)
     }
 
@@ -116,19 +116,19 @@ class ActivityDrugaGodina : AppCompatActivity() {
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        //val uloga = intent.getStringExtra("Uloga")
-        //return if ("Admin" == uloga) {
+        val uloga = intent.getStringExtra("Uloga")
+        return if ("Admin" == uloga) {
         val menuInflater = menuInflater
         menuInflater.inflate(R.menu.delete_all_kolegij_menu, menu)
         return true
-        //} else {
-        //    false
-        //}
+        } else {
+            false
+        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        //val uloga = intent.getStringExtra("Uloga")
-        //return if ("Admin" == uloga) {
+        val uloga = intent.getStringExtra("Uloga")
+        return if ("Admin" == uloga) {
         return when (item.itemId) {
             R.id.delete_all_kolegiji -> {
                 drugaGodinaViewModel?.deleteAll()
@@ -137,9 +137,9 @@ class ActivityDrugaGodina : AppCompatActivity() {
             }
             else -> super.onOptionsItemSelected(item)
         }
-        //} else {
-        //    false
-        //}
+        } else {
+            false
+        }
     }
 
     companion object {
